+++ src/CHANGELOG.md
@@ -8,9 +8,13 @@ v.3.9
 * [bug] Taxonomy pages did not work for people's names with accented letters (added sanitize_title())
 * [bug] Wrong annotation for class-settings.php (extra OPTIONS_WIDGET properties)
 * [bug] ImdbPHP search was not working for people on some sporadic circumstances
+* [bug] Creating Cache folder (method settings->lumiere_create_cache() ) was not properly updating path option in database.
+* [bug] If cache was not utilised, IMDB pictures were tried to be retrieved anyway (IMDBphp needs a folder to store and display pictures). Edited frontend classes (movie, popup-movie and popup-person) and theme taxonomy-people-standard
+* [bug] Uninstall did not remove the cache path as saved in database, but Lumiere\Settings::LUMIERE_CACHE_OPTIONS instead
 * [technical] Properly using composer scripts in src, not copied to the production vendor folder with composer anymore
 * [technical] Using composer autoload, removed all home-made autoloads. Can't use PSR-4 autoloading, since WordPress naming is not PSR-4 compliant.
 * [technical] Bootstrap Window Maker is now used through composer
+* [technical] Removed cache option 'imdbstorecache'; using cache option 'usecache' in class-imdbphp instead to define it
 * [technical] Updated to bootstrap 5.2.2
 * [technical] Added link to settings in WP plugins page
 * [technical] Updated to IMDBPHP library 7.3.1 fixing major bugs